+++ changelog.md
@@ -7,6 +7,7 @@
 - Add Mesos Label support to (as example) control traefik
 - Add possibility to run the K8 Server and Agent on a specified hostname
 - Fix status information of all m3s services
+- Mesos-CLI: Resolve framework ID also by name
 
 ## v0.1.0